requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
requests==2.31.0
yt-dlp>=2025.11.0
aiohttp>=3.9.0